Module rasslib::tree

Contains structs and traits for a Tree structure. Subtrees are of the same type Tree<T>. A visitor-pattern trait is provided which also internally used for the TreePrinter.

Path represents a path through the tree to a specific not. Finding an element returns a vector of paths.

Structs

Path
Represents a path to a specific element in the tree structure. Not that parameter T shall be of the same type as for the corresponding Tree<T>.

Tree
A Tree structure which contains elements that are also trees?

TreeIterator

TreePrinter

Traits

TreeVisitor
The TreeVisitor trait is the visitor part of the visitor pattern. Objects which on to get notified about a visited node while a tree structure is traversed this trait should be implemented.